refactor: separate out as DocumentExplorer for reuse
