+++ shared/i18n/locales/en_US/translation.json
@@ -113,6 +113,9 @@
   "Default collection": "Default collection",
   "Deleted Collection": "Deleted Collection",
   "Unpin": "Unpin",
+  "Search collections & documents": "Search collections & documents",
+  "No results found": "No results found",
+  "Untitled": "Untitled",
   "New": "New",
   "Only visible to you": "Only visible to you",
   "Draft": "Draft",
@@ -199,7 +202,6 @@
   "Click to retry": "Click to retry",
   "Back": "Back",
   "Documents": "Documents",
-  "Untitled": "Untitled",
   "Results": "Results",
   "No results for {{query}}": "No results for {{query}}",
   "Logo": "Logo",
@@ -508,13 +510,11 @@
   "Document moved": "Document moved",
   "Current location": "Current location",
   "Choose a new location": "Choose a new location",
-  "Search collections & documents": "Search collections & documents",
   "Couldn’t create the document, try again?": "Couldn’t create the document, try again?",
   "Document permanently deleted": "Document permanently deleted",
   "Are you sure you want to permanently delete the <em>{{ documentTitle }}</em> document? This action is immediate and cannot be undone.": "Are you sure you want to permanently delete the <em>{{ documentTitle }}</em> document? This action is immediate and cannot be undone.",
   "Select a location to publish": "Select a location to publish",
   "Couldn’t publish the document, try again?": "Couldn’t publish the document, try again?",
-  "No results found": "No results found",
   "Publish in <em>{{ location }}</em>": "Publish in <em>{{ location }}</em>",
   "view and edit access": "view and edit access",
   "view only access": "view only access",
